internal/util/{file,yaml}util: moved all yaml functions to yamlutil
