+++ test/e2e/memcached_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/ghodss/yaml"
 	"github.com/operator-framework/operator-sdk/internal/util/fileutil"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
+	"github.com/operator-framework/operator-sdk/internal/util/yamlutil"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	"github.com/operator-framework/operator-sdk/pkg/test/e2eutil"
 
@@ -222,7 +223,7 @@ func TestMemcached(t *testing.T) {
 			"vendor/github.com/operator-framework/operator-sdk/pkg")
 	}
 
-	file, err := projutil.GenerateCombinedGlobalManifest()
+	file, err := yamlutil.GenerateCombinedGlobalManifest()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -457,7 +458,7 @@ func MemcachedCluster(t *testing.T) {
 		}
 	}
 
-	file, err := projutil.GenerateCombinedNamespacedManifest()
+	file, err := yamlutil.GenerateCombinedNamespacedManifest()
 	if err != nil {
 		t.Fatal(err)
 	}
